Crate dashboard

dashboard

dashboard is list of “GitHub Actions build status” of all the repos under my account, for my own use.

Usage

Add this to your markdown file:

<!--START_SECTION:dashboard-->
<!--END_SECTION:dashboard-->

try_main_refactor_v3 function serves as the main driver of the program:

It is responsible for creating and initializing an instance of App and using it to call GitCliOps::fetch_repos_write_data to fetch the list of Git repositories and write the data to disk.

Then, it spawns two operations (update_markdown_file and write_json_file) into separate threads for parallel execution.

Examples

use crate::dashboard::app::*;

#[tokio::main]
pub async fn main() -> Result<(), AppError> {
    dotenv::dotenv().ok();
    let file_path = "/path/to/README.md";
    let result = try_main_refactor_v3(file_path).await;

    match result {
        Ok(()) => println!("Success"),
        Err(e) => println!("Error: {:?}", e),
    }
    Ok(())
}

Error

This function will return an Err variant of Result type if the call to GitCliOps::fetch_repos_write_data fails or if the thread::scope call returns an error. In this case, AppError will be returned.

Development

Usage

$ CARGO_LOG=error cargo r -p dashboard

Data - Github API

Output of Github CLI command gh repo list is serialized to gh_repo_list.json

Parsed API - Markdown

The detail of each git repository is appended as a list item to README.md:

* [name](url) — description

name - Repository name

url - Repository URL

description - Description of the repository

Modules

app

app module contains App which contains prelude for all modules in this crate.