Function dashboard::app::try_main_refactor_v3

pub fn try_main_refactor_v3(file_path: &str) -> Result<(), AppError>

try_main_refactor_v3 function serves as the main driver of the program.

It is responsible for creating and initializing an instance of App and using it to call GitCliOps::fetch_repos_write_data to fetch the list of Git repositories and write the data to disk. Then, it spawns two operations (update_markdown_file and write_json_file) into separate threads for parallel execution.

Error

This function will return an Err variant of Result type if the call to GitCliOps::fetch_repos_write_data fails or if the thread::scope call returns an error. In this case, AppError will be returned.

Examples

use crate::dashboard::app::*;

let file_path = "/path/to/README.md";
let result = try_main_refactor_v3(file_path);

match result {
    Ok(()) => println!("Success"),
    Err(e) => println!("Error: {:?}", e),
}