Dashboard

Module eco_can

Module for CAN communication

CAN Package Information

A CAN package is setup like this:

#[allow(non_camel_case_types)]
#[derive(bincode::Encode, bincode::Decode, PartialEq, Clone, Debug)]
#[repr(C)]
pub struct FDCAN_PACKAGE_NAME {
    // Package Data
}
impl FDCANPack for FDCAN_FetPack_t {
   const FDCAN_BYTES: u8 = BYTE_LENGTH; // set this to the size of the package in bytes
   const FDCAN_ID: u16 = CAN_ID;    // the ID of the CAN package
}

#[allow(non_camel_case_types)] allows non-camel-case names for FDCAN packages

#[derive(bincode::Encode, bincode::Decode)] makes the package able to be encoded to and decoded from bytes.

#[derive(Clone)] allows the package to be copied.

#[derive(Debug)] allows the package to be formatted in log messages.

#[repr(C)] Make Rust use the same memory layout for this struct as C to ensure compatility. For more information: https://doc.rust-lang.org/nomicon/other-reprs.html

Structs

ECOCAN_H2Pack1_t

ECOCAN_H2Pack2_t

ECOCAN_H2_ARM_ALARM_t

ECOCAN_RelPackChrg_t

FDCAN_BATTPack2_t

FDCAN_BOOSTPack2_t

FDCAN_BOOSTPack3_t

FDCAN_BOOSTPack_t

FDCAN_FccPack1_t

FDCAN_FccPack2_t

FDCAN_FccPack3_t

FDCAN_FetPack_t

FDCAN_RelPackCap_t

FDCAN_RelPackFc_t

FDCAN_RelPackMtr_t

FDCAN_RelPackNrg_t

Constants

FDCAN_H2ALARM_ID

1 indicates tripped alarm

FDCAN_RELSTATE_ID

Relay State ID

FDCAN_SYNCLED_ID

1 indicates led on

Traits

FDCANPack

Prerequisite trait for FDCAN Packages