Trait dashboard::gh::GitCliOps

pub trait GitCliOps {
    fn fetch_repos_write_data(&mut self) -> Result<(), AppError>;
}

Required Methods

fn fetch_repos_write_data(&mut self) -> Result<(), AppError>

Use GitHub CLI gh utility in xshell to fetch list of repositories and, mutate self.data to the json response of Vec<GitRepo>.

xshell::Shell - doesn’t use the shell directly, but rather re-implements parts of scripting environment in Rust.

Errors

This function will return an error if:

std::env::current_dir - returns an error while creating new [xshell::Shell].

[xshell::cmd!] - on read returns a non-zero return code considered to be an error.

serde_json - conversion can fail if the structure of the input does not match the structure expected by Vec<GitRepo>.