Module dashboard::markdown

Structs

NEW_CONTENT 🔒

Enums

SectionState 🔒

Indicates if visitor is inside markdown start and end section blocks.

Constants

DESC_WC

Word count limit for description.

Statics

LAZY 🔒

Functions

__stability 🔒

__static_ref_initialize 🔒

fmt_markdown_list_item 🔒

Create and format a new markdown list item with repo name, url and its description.

main 🔒

update_markdown_file 🔒

This way, the function returns a Result which can be checked by the caller for success or failure. The error type is Box, which allows for the caller to handle any type of error that may occur during the execution of the function. The code is more concise and easier to read, with the use of a macro to simplify the comment block strings.