Crate dashboard

dashboard

dashboard is list of “GitHub Actions build status” of all the repos under my account, for my own use.

Usage

Add this to your markdown file:

<!--START_SECTION:dashboard-->
<!--END_SECTION:dashboard-->

Development

Usage

$ CARGO_LOG=error cargo r -p dashboard

Data - Github API

Output of Github CLI command gh repo list is serialized to gh_repo_list.json

Parsed API - Markdown

The detail of each git repository is appended as a list item to README.md:

* [name](url) — description

name - Repository name

url - Repository URL

description - Description of the repository

Structs

App

Enums

AppError

AppError Instead of cloning the std::io::Error, we can store the error within the AppError as an Arc (Atomic Reference Counted) smart pointer. Allows for multiple references to the same error to be stored in different places without having to clone it.

Constants

DESC_WC

Word count limit for description.

Functions

main

main entrypoint.

try_main_refactor

Type Definitions

Result

Result<T, E>