Crate dashboard

dashboard

dashboard is list of “GitHub Actions build status” of all the repos under my account, for my own use.

Development

Usage

$ CARGO_LOG=error cargo r -p dashboard

Data - Github API

Output of Github CLI command gh repo list is serialized to gh_repo_list.json

Parsed API - Markdown

The detail of each git repository is appended as a list item to README.md:

* [name](url) — description

name - Repository name

url - Repository URL

description - Description of the repository

Re-exports

pub use crate::app::AppError;

pub use crate::app::Result;

pub use crate::gh::GitCliOps;

pub use crate::gh::GitRepo;

pub use crate::gh::RepositoryTopic;

Modules

app

app module contains App which contains prelude for all modules in this crate.

config

db

gh

Functions

main

try_main